got sesion request working

+++ f1data/services/data_transformation/integers.py
@@ -1,5 +1,4 @@
 import pandas as pd
-import numpy as np
 
 
-int_or_null = lambda x: np.int16(x) if not pd.isna(x) else None
+int_or_null = lambda x: int(x) if not pd.isna(x) else None
