+++ f1data/services/data_transformation/speed.py
@@ -1,5 +1,3 @@
-import numpy as np
 import pandas as pd
 
-
-speed_to_float_or_null = lambda x: np.float16(x) if not pd.isna(x) else None
+speed_to_float_or_null = lambda x: float(x) if not pd.isna(x) else None
